restored find_center.py as independent module
